+++ src/Scheduler.h
@@ -21,7 +21,7 @@ public:
 	virtual void pushInst(CInstrument* pInstToPush, float fDurationInSec = 1.0f, float fOnsetInSec = 0.0f);
 
 	// Returns schedule length in samples
-	int getLength() const;
+	int64_t getLength() const;
 
 	// Frame-by-Frame processing function
 	virtual void process(float** ppfOutBuffer, int iNumChannels, int iCurrentFrame) override;
@@ -36,8 +36,8 @@ protected:
 	unordered_set<CSoundProcessor*> m_GarbageCollector;
 
 	// This can be viewed as the schedule's internal clock
-	long long m_iSampleCounter = 0;
-	int m_iScheduleLength = 0;
+	int64_t m_iSampleCounter = 0;
+	int64_t m_iScheduleLength = 0;
 
 	// Extra buffer space for applying adsr and gain
 	const int m_iMaxChannels = 6;
@@ -46,9 +46,9 @@ protected:
 	// These maps are used to trigger events on instruments at the correct sample
 	// Key: Current Frame
 	// Value: Set of Instruments with an event at that key
-	map<int, unordered_set<CInstrument*>> m_MapNoteOn;
-	map<int, unordered_set<CInstrument*>> m_MapNoteOff;
-	map<int, unordered_set<CInstrument*>> m_MapRemover;
+	map<int64_t, unordered_set<CInstrument*>> m_MapNoteOn;
+	map<int64_t, unordered_set<CInstrument*>> m_MapNoteOff;
+	map<int64_t, unordered_set<CInstrument*>> m_MapRemover;
 
 	// Helper function to check if there is a trigger at a specified sample
 	// Will return the set of instruments pertaining to the trigger if so
